feat(#44): add separate logger for progressbar
